Fix not dying upon respawn cancel
This occurs due to not having low enough hp for death to occur. So
whether `kill_now` flag is set or not wouldn't matter, but at the same
time, the hp is below 1 hp which triggers our logic to cessate the
player and show the UI, and so we get stuck in a loop depending on what
the hp of the player was the start of the cycle.